Address global search review lifecycle feedback
