+++ Sources/Search/GlobalSearchCoordinator.swift
@@ -2,31 +2,17 @@ import AppKit
 import Foundation
 
 @MainActor
-struct GlobalSearchPanelContext {
-    let windowID: UUID
-    let windowTitle: String
-    let workspaceID: UUID
-    let workspaceTitle: String
-    let panelID: UUID
-    let panelTitle: String
-    let panel: any Panel
-
-    var location: String {
-        "\(windowTitle) > \(workspaceTitle)"
-    }
-}
-
-@MainActor
 final class GlobalSearchCoordinator {
     static let shared = GlobalSearchCoordinator()
 
-    private let maxIndexedTextCharacters = 400_000
     private let browserCaptureDebounceMilliseconds = 250
     private let markdownCaptureDebounceMilliseconds = 250
     private var browserCaptureTasks: [UUID: Task<Void, Never>] = [:]
     private var browserCaptureTaskIDs: [UUID: UUID] = [:]
     private var markdownCaptureTasks: [UUID: Task<Void, Never>] = [:]
     private var markdownCaptureTaskIDs: [UUID: UUID] = [:]
+    private var panelPurgeTasks: [UUID: Task<Void, Never>] = [:]
+    private var panelPurgeTaskIDs: [UUID: UUID] = [:]
     private var startupIndexTask: Task<Void, Never>?
     private var indexState: SearchIndexState = .idle
     private lazy var popover = MenubarSearchPopover(coordinator: self)
@@ -87,8 +73,9 @@ final class GlobalSearchCoordinator {
 
         for context in appDelegate.globalSearchPanelContexts() {
             guard !Task.isCancelled else { return }
+            cancelPanelPurge(forPanelID: context.panelID)
 
-            let titleDocument = titleDocument(for: context)
+            let titleDocument = GlobalSearchDocuments.titleDocument(for: context)
             do {
                 try await index.upsert(titleDocument)
             } catch {
@@ -103,7 +90,7 @@ final class GlobalSearchCoordinator {
                 if markdownPanel.isFileUnavailable {
                     cancelMarkdownCapture(forPanelID: context.panelID)
                     await purgeMarkdownDocument(forPanelID: context.panelID, index: index)
-                } else if let document = markdownDocument(for: markdownPanel, context: context) {
+                } else if let document = GlobalSearchDocuments.markdownDocument(for: markdownPanel, context: context) {
                     do {
                         try await index.upsert(document)
                     } catch {
@@ -121,6 +108,7 @@ final class GlobalSearchCoordinator {
     func captureBrowserPanel(_ panel: BrowserPanel) {
         let panelID = panel.id
         let taskID = UUID()
+        cancelPanelPurge(forPanelID: panelID)
         browserCaptureTasks[panelID]?.cancel()
         browserCaptureTaskIDs[panelID] = taskID
 
@@ -154,18 +142,35 @@ final class GlobalSearchCoordinator {
         let panelID = panel.id
         guard !panel.isFileUnavailable else {
             cancelMarkdownCapture(forPanelID: panelID)
-            Task { @MainActor [weak self] in
-                guard let self, let index = await self.ensureIndex() else { return }
+            let taskID = UUID()
+            markdownCaptureTaskIDs[panelID] = taskID
+            let task = Task { @MainActor [weak self] in
+                guard let self else { return }
+                defer {
+                    if self.markdownCaptureTaskIDs[panelID] == taskID {
+                        self.markdownCaptureTasks[panelID] = nil
+                        self.markdownCaptureTaskIDs[panelID] = nil
+                    }
+                }
+
+                guard !Task.isCancelled,
+                      self.markdownCaptureTaskIDs[panelID] == taskID,
+                      let index = await self.ensureIndex() else {
+                    return
+                }
+
                 await self.purgeMarkdownDocument(forPanelID: panelID, index: index)
             }
+            markdownCaptureTasks[panelID] = task
             return
         }
 
+        cancelPanelPurge(forPanelID: panelID)
         guard let context = AppDelegate.shared?.globalSearchContext(
             forPanelID: panel.id,
             preferredWorkspaceID: panel.workspaceId
         ),
-            let document = markdownDocument(for: panel, context: context) else {
+            let document = GlobalSearchDocuments.markdownDocument(for: panel, context: context) else {
             return
         }
 
@@ -210,16 +215,42 @@ final class GlobalSearchCoordinator {
         browserCaptureTasks[panelID] = nil
         browserCaptureTaskIDs[panelID] = nil
         cancelMarkdownCapture(forPanelID: panelID)
-        Task { @MainActor [weak self] in
-            guard let self, let index = await self.ensureIndex() else { return }
+        panelPurgeTasks[panelID]?.cancel()
+
+        let taskID = UUID()
+        panelPurgeTaskIDs[panelID] = taskID
+        let task = Task { @MainActor [weak self] in
+            guard let self else { return }
+            defer {
+                if self.panelPurgeTaskIDs[panelID] == taskID {
+                    self.panelPurgeTasks[panelID] = nil
+                    self.panelPurgeTaskIDs[panelID] = nil
+                }
+            }
+
+            guard !Task.isCancelled,
+                  self.panelPurgeTaskIDs[panelID] == taskID,
+                  let index = await self.ensureIndex() else {
+                return
+            }
+
             do {
+                guard !Task.isCancelled, self.panelPurgeTaskIDs[panelID] == taskID else { return }
                 try await index.deletePanel(panelID)
             } catch {
+                guard !Task.isCancelled else { return }
 #if DEBUG
                 cmuxDebugLog("globalSearch.panel.purge failed panel=\(panelID.uuidString.prefix(5)) error=\(error.localizedDescription)")
 #endif
             }
         }
+        panelPurgeTasks[panelID] = task
+    }
+
+    private func cancelPanelPurge(forPanelID panelID: UUID) {
+        panelPurgeTasks[panelID]?.cancel()
+        panelPurgeTasks[panelID] = nil
+        panelPurgeTaskIDs[panelID] = nil
     }
 
     private func cancelMarkdownCapture(forPanelID panelID: UUID) {
@@ -285,14 +316,14 @@ final class GlobalSearchCoordinator {
         let payload = await browserPagePayload(for: panel)
         guard !Task.isCancelled else { return }
         let fallbackTitle = panel.displayTitle.trimmingCharacters(in: .whitespacesAndNewlines)
-        let title = firstNonEmpty(payload?.title, panel.pageTitle, fallbackTitle)
+        let title = GlobalSearchDocuments.firstNonEmpty(payload?.title, panel.pageTitle, fallbackTitle)
             ?? String(localized: "globalSearch.untitled", defaultValue: "Untitled")
-        let location = firstNonEmpty(payload?.url, panel.currentURL?.absoluteString) ?? ""
-        let bodyText = firstNonEmpty(payload?.text) ?? ""
-        let text = cappedText([title, location, bodyText].filter { !$0.isEmpty }.joined(separator: "\n"))
+        let location = GlobalSearchDocuments.firstNonEmpty(payload?.url, panel.currentURL?.absoluteString) ?? ""
+        let bodyText = GlobalSearchDocuments.firstNonEmpty(payload?.text) ?? ""
+        let text = GlobalSearchDocuments.cappedText([title, location, bodyText].filter { !$0.isEmpty }.joined(separator: "\n"))
         guard !text.isEmpty else { return }
 
-        let anchor = firstNonEmpty(location, panel.id.uuidString) ?? panel.id.uuidString
+        let anchor = GlobalSearchDocuments.firstNonEmpty(location, panel.id.uuidString) ?? panel.id.uuidString
         let document = SearchIndexDocument(
             id: SearchIndexDocument.panelStableID(panelID: context.panelID, kind: .browser),
             windowID: context.windowID,
@@ -319,7 +350,7 @@ final class GlobalSearchCoordinator {
     private func browserPagePayload(for panel: BrowserPanel) async -> BrowserPagePayload? {
         let script = """
         (() => {
-            const limit = \(maxIndexedTextCharacters);
+            const limit = \(GlobalSearchIndexingLimits.maxIndexedTextCharacters);
             const collectText = (root) => {
                 if (!root) { return ""; }
                 const walker = document.createTreeWalker(root, NodeFilter.SHOW_TEXT);
@@ -353,55 +384,6 @@ final class GlobalSearchCoordinator {
         }
     }
 
-    private func titleDocument(for context: GlobalSearchPanelContext) -> SearchIndexDocument {
-        let text = [
-            context.windowTitle,
-            context.workspaceTitle,
-            context.panelTitle
-        ].filter { !$0.isEmpty }.joined(separator: "\n")
-
-        return SearchIndexDocument(
-            id: SearchIndexDocument.panelStableID(panelID: context.panelID, kind: .title),
-            windowID: context.windowID,
-            workspaceID: context.workspaceID,
-            panelID: context.panelID,
-            kind: .title,
-            title: context.panelTitle,
-            location: context.location,
-            anchor: "title",
-            text: text
-        )
-    }
-
-    private func markdownDocument(for panel: MarkdownPanel, context: GlobalSearchPanelContext) -> SearchIndexDocument? {
-        let title = panel.displayTitle.trimmingCharacters(in: .whitespacesAndNewlines)
-        let text = cappedText([title, panel.filePath, panel.content].filter { !$0.isEmpty }.joined(separator: "\n"))
-        guard !text.isEmpty else { return nil }
-
-        return SearchIndexDocument(
-            id: SearchIndexDocument.panelStableID(panelID: context.panelID, kind: .markdown),
-            windowID: context.windowID,
-            workspaceID: context.workspaceID,
-            panelID: context.panelID,
-            kind: .markdown,
-            title: title,
-            location: panel.filePath,
-            anchor: panel.filePath,
-            text: text
-        )
-    }
-
-    private func cappedText(_ text: String) -> String {
-        guard text.count > maxIndexedTextCharacters else { return text }
-        let endIndex = text.index(text.startIndex, offsetBy: maxIndexedTextCharacters)
-        return String(text[..<endIndex])
-    }
-
-    private func firstNonEmpty(_ values: String?...) -> String? {
-        values
-            .compactMap { $0?.trimmingCharacters(in: .whitespacesAndNewlines) }
-            .first { !$0.isEmpty }
-    }
 }
 
 private enum SearchIndexState {
@@ -409,188 +391,4 @@ private enum SearchIndexState {
     case opening(Task<SearchIndex, Error>)
     case ready(SearchIndex)
     case failed
-}
-
-private struct BrowserPagePayload: Decodable {
-    let title: String
-    let url: String
-    let text: String
-}
-
-extension AppDelegate {
-    func globalSearchPanelContexts() -> [GlobalSearchPanelContext] {
-        var contexts: [GlobalSearchPanelContext] = []
-        var seenPanelKeys = Set<String>()
-        var windowOrdinal = 1
-
-        func append(windowID: UUID, tabManager: TabManager, window: NSWindow?) {
-            let fallbackWindowTitle = String(localized: "menu.windowNumber", defaultValue: "Window \(windowOrdinal)")
-            windowOrdinal += 1
-            let windowTitle = {
-                let trimmed = window?.title.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-                return trimmed.isEmpty ? fallbackWindowTitle : trimmed
-            }()
-
-            for workspace in tabManager.tabs {
-                let workspaceTitle = {
-                    let trimmed = workspace.title.trimmingCharacters(in: .whitespacesAndNewlines)
-                    return trimmed.isEmpty
-                        ? String(localized: "workspace.displayName.fallback", defaultValue: "Workspace")
-                        : trimmed
-                }()
-
-                for panel in workspace.panels.values {
-                    let key = "\(windowID.uuidString):\(workspace.id.uuidString):\(panel.id.uuidString)"
-                    guard seenPanelKeys.insert(key).inserted else { continue }
-                    let panelTitle = panel.displayTitle.trimmingCharacters(in: .whitespacesAndNewlines)
-                    contexts.append(
-                        GlobalSearchPanelContext(
-                            windowID: windowID,
-                            windowTitle: windowTitle,
-                            workspaceID: workspace.id,
-                            workspaceTitle: workspaceTitle,
-                            panelID: panel.id,
-                            panelTitle: panelTitle.isEmpty ? workspaceTitle : panelTitle,
-                            panel: panel
-                        )
-                    )
-                }
-            }
-        }
-
-        for context in mainWindowContexts.values {
-            append(
-                windowID: context.windowId,
-                tabManager: context.tabManager,
-                window: context.window ?? windowForMainWindowId(context.windowId)
-            )
-        }
-
-        for route in recoverableMainWindowRoutes() {
-            guard let tabManager = route.tabManager else { continue }
-            let alreadySeen = contexts.contains { $0.windowID == route.windowId }
-            guard !alreadySeen else { continue }
-            append(windowID: route.windowId, tabManager: tabManager, window: route.window)
-        }
-
-        return contexts
-    }
-
-    func globalSearchContext(
-        forPanelID panelID: UUID,
-        preferredWorkspaceID: UUID?
-    ) -> GlobalSearchPanelContext? {
-        var fallback: GlobalSearchPanelContext?
-        var seenWindowIDs = Set<UUID>()
-        var windowOrdinal = 1
-
-        func inspect(windowID: UUID, tabManager: TabManager, window: NSWindow?) -> GlobalSearchPanelContext? {
-            _ = seenWindowIDs.insert(windowID)
-            let fallbackWindowTitle = String(localized: "menu.windowNumber", defaultValue: "Window \(windowOrdinal)")
-            windowOrdinal += 1
-            let windowTitle = {
-                let trimmed = window?.title.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-                return trimmed.isEmpty ? fallbackWindowTitle : trimmed
-            }()
-
-            for workspace in tabManager.tabs {
-                guard let panel = workspace.panels[panelID] else { continue }
-                let workspaceTitle = {
-                    let trimmed = workspace.title.trimmingCharacters(in: .whitespacesAndNewlines)
-                    return trimmed.isEmpty
-                        ? String(localized: "workspace.displayName.fallback", defaultValue: "Workspace")
-                        : trimmed
-                }()
-                let panelTitle = panel.displayTitle.trimmingCharacters(in: .whitespacesAndNewlines)
-                let context = GlobalSearchPanelContext(
-                    windowID: windowID,
-                    windowTitle: windowTitle,
-                    workspaceID: workspace.id,
-                    workspaceTitle: workspaceTitle,
-                    panelID: panel.id,
-                    panelTitle: panelTitle.isEmpty ? workspaceTitle : panelTitle,
-                    panel: panel
-                )
-
-                if preferredWorkspaceID == nil || workspace.id == preferredWorkspaceID {
-                    return context
-                }
-                fallback = fallback ?? context
-            }
-
-            return nil
-        }
-
-        for context in mainWindowContexts.values {
-            if let result = inspect(
-                windowID: context.windowId,
-                tabManager: context.tabManager,
-                window: context.window ?? windowForMainWindowId(context.windowId)
-            ) {
-                return result
-            }
-        }
-
-        for route in recoverableMainWindowRoutes() {
-            guard !seenWindowIDs.contains(route.windowId),
-                  let tabManager = route.tabManager else {
-                continue
-            }
-            if let result = inspect(windowID: route.windowId, tabManager: tabManager, window: route.window) {
-                return result
-            }
-        }
-
-        return fallback
-    }
-
-    func openGlobalSearchHit(_ hit: SearchIndexHit, query: String) {
-        let resolvedContext = hit.panelID.flatMap {
-            globalSearchContext(forPanelID: $0, preferredWorkspaceID: hit.workspaceID)
-        }
-        let windowID = resolvedContext?.windowID ?? hit.windowID
-        let workspaceID = resolvedContext?.workspaceID ?? hit.workspaceID
-
-        guard let tabManager = tabManagerFor(windowId: windowID),
-              let workspace = tabManager.tabs.first(where: { $0.id == workspaceID }) else {
-            NSSound.beep()
-            return
-        }
-
-        _ = focusMainWindow(windowId: windowID)
-        tabManager.selectTab(workspace)
-        TerminalController.shared.setActiveTabManager(tabManager)
-
-        if let panelID = hit.panelID, workspace.panels[panelID] != nil {
-            tabManager.focusSurface(tabId: workspace.id, surfaceId: panelID)
-            if let browserPanel = workspace.browserPanel(for: panelID) {
-                applyBrowserInlineSearch(query: query, hit: hit, to: browserPanel)
-            }
-        }
-    }
-
-    private func applyBrowserInlineSearch(query: String, hit: SearchIndexHit, to panel: BrowserPanel) {
-        guard let needle = GlobalSearchInlineSearch.browserNeedle(for: query, hit: hit) else { return }
-        if let searchState = panel.searchState {
-            searchState.needle = needle
-        } else {
-            panel.searchState = BrowserSearchState(needle: needle)
-        }
-    }
-}
-
-enum GlobalSearchInlineSearch {
-    static func browserNeedle(for query: String, hit: SearchIndexHit) -> String? {
-        let tokens = SearchIndex.queryTokens(for: query)
-        guard !tokens.isEmpty else { return nil }
-
-        let hitText = [
-            hit.snippet,
-            hit.title,
-            hit.location,
-            hit.anchor
-        ].joined(separator: "\n").lowercased()
-
-        return tokens.first { hitText.contains($0) } ?? tokens[0]
-    }
 }
